refactor : 로또 번호 생성하는 함수 CreateLottos 클래스에 포함
